refactor(core): extract addon runtime into addon-runner, tighten overrides
Address PR #924 review:
- Move addon singleton-service resolution and instance helpers out of
  function-runner.ts and rpc-runner.ts into wirings/rpc/addon-runner.ts,
  keeping function-runner (the core execution path) free of addon logic.
- Shorten the oversized cache-key comment.
- Fix wire-addon test override maps to alias names→names (not values).
